Rename `_ExclusiveSpatialAudioAttacher` class to `_SpatialAudioAttacher`
